draft kconfig setup and log system
