+++ prj.conf
@@ -2,3 +2,9 @@ CONFIG_PRINTK=y
 CONFIG_HEAP_MEM_POOL_SIZE=256
 CONFIG_ASSERT=y
 CONFIG_GPIO=y
+
+# Enable logging
+CONFIG_LOG=y
+
+# Set default log level
+CONFIG_LOG_DEFAULT_LEVEL=3
