+++ services/Kconfig.services
@@ -1,0 +1,28 @@
+#menu "services options"
+
+config IOMICO_SERVICES
+    bool "service options"
+    default n
+
+if  IOMICO_SERVICES
+
+config IOMICO_SERVICES_A
+    int "option services number A" 
+    default 123
+    depends on IOMICO_OPTION_A
+
+config IOMICO_SERVICES_B
+    hex "option services number B" 
+    default 123
+
+config IOMICO_SERVICES_C
+    hex "option services number C" 
+    default 1 if IOMICO_OPTION_A && IOMICO_OPTION_B > 10
+
+config IOMICO_SERVICES_D
+    bool "option services number D" 
+    select IOMICO_OPTION_A #forces the selection of IOMICO_OPTION_A
+
+endif 
+
+#endmenu
