+++ src/main.c
@@ -1,180 +1,13 @@
-/*
- * Copyright (c) 2016 Intel Corporation
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
 #include <stdio.h>
-#include <zephyr/kernel.h>
-#include <zephyr/drivers/gpio.h>
-#include <zephyr/device.h>
-#include <zephyr/drivers/gpio.h>
-#include <zephyr/sys/printk.h>
-#include <zephyr/sys/__assert.h>
-#include <string.h>
-#include <zephyr/sys/util.h>
-#include <inttypes.h>
+#include <zephyr/logging/log.h>
 
-/* 1000 msec = 1 sec */
-#define SLEEP_TIME_MS   200
+LOG_MODULE_REGISTER(my_module, LOG_LEVEL_DBG);
 
-/* size of stack area used by each thread */
-#define STACKSIZE 1024
-
-/* scheduling priority used by each thread */
-#define PRIORITY 7
-
-/* The devicetree node identifier for the "led0" alias. */
-#define LED0_NODE DT_ALIAS(led0)
-#define LED1_NODE DT_ALIAS(led1)
-
-/*
- * A build error on this line means your board is unsupported.
- * See the sample documentation for information on how to fix this.
- */
-static const struct gpio_dt_spec led = GPIO_DT_SPEC_GET(LED0_NODE, gpios);
-
-/*
- * Get button configuration from the devicetree sw0 alias. This is mandatory.
- */
-#define SW0_NODE	DT_ALIAS(sw0)
-#if !DT_NODE_HAS_STATUS_OKAY(SW0_NODE)
-#error "Unsupported board: sw0 devicetree alias is not defined"
-#endif
-static const struct gpio_dt_spec button = GPIO_DT_SPEC_GET_OR(SW0_NODE, gpios, {0});
-static struct gpio_callback button_cb_data;
-
-struct printk_data_t {
-	void *fifo_reserved; /* 1st word reserved for use by fifo */
-	uint32_t led;
-	uint32_t cnt;
-};
-
-K_FIFO_DEFINE(printk_fifo);
-
-struct led {
-	struct gpio_dt_spec spec;
-	uint8_t num;
-};
-
-static const struct led led0 = {
-	.spec = GPIO_DT_SPEC_GET_OR(LED0_NODE, gpios, {0}),
-	.num = 0,
-};
-
-static const struct led led1 = {
-	.spec = GPIO_DT_SPEC_GET_OR(LED1_NODE, gpios, {0}),
-	.num = 1,
-};
-
-void blink(const struct led *led, uint32_t sleep_ms, uint32_t id)
+int main() 
 {
-	const struct gpio_dt_spec *spec = &led->spec;
-	int cnt = 0;
-	int ret;
+   LOG_INF("hola");
+   LOG_ERR("hola");
+// printf("Hello, World! %u %u\n",CONFIG_IOMICO_OPTION_A,CONFIG_IOMICO_OPTION_B);
 
-	if (!device_is_ready(spec->port)) {
-		printk("Error: %s device is not ready\n", spec->port->name);
-		return;
-	}
-
-	ret = gpio_pin_configure_dt(spec, GPIO_OUTPUT);
-	if (ret != 0) {
-		printk("Error %d: failed to configure pin %d (LED '%d')\n",
-			ret, spec->pin, led->num);
-		return;
-	}
-
-	while (1) {
-		gpio_pin_set(spec->port, spec->pin, cnt % 2);
-
-		struct printk_data_t tx_data = { .led = id, .cnt = cnt };
-
-		size_t size = sizeof(struct printk_data_t);
-		char *mem_ptr = k_malloc(size);
-		__ASSERT_NO_MSG(mem_ptr != 0);
-
-		memcpy(mem_ptr, &tx_data, size);
-
-		k_fifo_put(&printk_fifo, mem_ptr);
-
-		k_msleep(sleep_ms);
-		cnt++;
-	}
+ return 0;
 }
-
-void blink0(void)
-{
-	blink(&led0, 100, 0);
-}
-
-void blink1(void)
-{
-	blink(&led1, 1000, 1);
-}
-
-void uart_out(void)
-{
-	while (1) {
-		struct printk_data_t *rx_data = k_fifo_get(&printk_fifo,
-							   K_FOREVER);
-		printk("Toggled led%d; counter=%d\n",
-		       rx_data->led, rx_data->cnt);
-		k_free(rx_data);
-	}
-}
-
-K_THREAD_DEFINE(blink0_id, STACKSIZE, blink0, NULL, NULL, NULL,
-		PRIORITY, 0, 0);
-K_THREAD_DEFINE(blink1_id, STACKSIZE, blink1, NULL, NULL, NULL,
-		PRIORITY, 0, 0);
-K_THREAD_DEFINE(uart_out_id, STACKSIZE, uart_out, NULL, NULL, NULL,
-		PRIORITY, 0, 0);
-
-
-void button_pressed(const struct device *dev, struct gpio_callback *cb, uint32_t pins)
-{
-	printk("Button pressed at %" PRIu32 "\n", k_cycle_get_32());
-}
-
-int btn(void)
-{
-	int ret;
-
-	if (!gpio_is_ready_dt(&button)) {
-		printk("Error: button device %s is not ready\n",
-		       button.port->name);
-		return 0;
-	}
-
-	ret = gpio_pin_configure_dt(&button, GPIO_INPUT);
-	if (ret != 0) {
-		printk("Error %d: failed to configure %s pin %d\n",
-		       ret, button.port->name, button.pin);
-		return 0;
-	}
-
-	ret = gpio_pin_interrupt_configure_dt(&button, GPIO_INT_EDGE_TO_ACTIVE);
-	if (ret != 0) {
-		printk("Error %d: failed to configure interrupt on %s pin %d\n",
-			ret, button.port->name, button.pin);
-		return 0;
-	}
-
-	gpio_init_callback(&button_cb_data, button_pressed, BIT(button.pin));
-	gpio_add_callback(button.port, &button_cb_data);
-	printk("Set up button at %s pin %d\n", button.port->name, button.pin);
-
-	printk("Press the button\n");
-    while (1) {
-       int val = gpio_pin_get_dt(&button);
-
-       if (val > 0) {
-          printk("btn pressed val: %u\n", val);
-       }
-       k_msleep(SLEEP_TIME_MS);
-    }
-	return 0;
-}
-
-K_THREAD_DEFINE(btn_id, STACKSIZE, btn, NULL, NULL, NULL, PRIORITY, 0, 0);
